prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,9 +7,9 @@
   *
   * Return: 0.
   */
-int main(long int  fib)
+int main(int fib)
 {
-	long int  fib1 = 1, fib2 = 1, swap; 
+	int  fib1 = 1, fib2 = 1, swap; 
 
 
 	for (fib = 1; fib <= 98; fib++)
